refactor(node): simplify node controller's code
